requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-ckeditor>=6.7.0
django-taggit>=5.0.0
django-mptt>=0.16.0
Pillow>=10.0.0
django-filter>=23.0
markdown>=3.5
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
django-simple-history>=3.4.0
